[DM-27044] First attempt at building hub image
Start off from the existing published jupyterhub image, and build
on top of that, rather than trying to install everything for now.

Since it isn't installed in a virtual environment, we'll have to
get rid of those parts, but that does simplify a few things since
we aren't actually running a service here.

In order to get an entrypoint into the nublado2 library, add
the jupyterhub_config.py which just calls into the library to
configure it.