[DM-27044] Changes for nublado2 chart
These are changes to be useful for lsst-sqre/charts#173 .

First, don't use jupyterhub_config.py, we'll use the chart, and I've
put the right incantation as the default in the chart, but also
have it here in the dev-values.yaml just to make it obvious and easy
to override and test, since I'm getting rid of the unused python
file.

Also, the stock zero 2 jupyterhub has enabled the idleculler and
I think that's fine for now, but it was struggling with that not
being installed in our image.  This updates the depedencies and
adds the idleculler in to fix the error message in the logs.