[DM-27245] Refactor config
Okay, so here I'm refactoring the config classes a bit.  I'm taking
out the nublado parts of the config from config.py, restoring that
to what it was.  I'm taking those contents and splitting it up into
the hub_config part (which will only be used on startup and is the
duct tape of this library) and the nublado_config part which will
be used by other modules in this repo a lot for config from the
yaml file.

I'm planning on changing the name of the yaml from hub_config.yaml
to nublado_config.yaml to be consistent as well.