[DM-27245] Arbitrary resource creation
In the dev-values.yaml there's an example of a namespace of the user
getting created after filling in the name and sending it to k8s.

On stop, we also clean it up.  This is just an example of how we
can create resources with just the yaml, and not call specific APIs
based on different resource types.