Move jupyterhub and tornado dependencies to dev
Avoid including JupyterHub or Tornado in the main dependencies, even
though the authenticator does depend on them, because we want to use
whatever versions were preinstalled in the Docker image that we use
as a base.  Move them to the dev dependencies instead so that the test
suite still works.