[DM-27278] Get gafaelfawr auth working on minikube
So the short version of this long story is that you can't access
minikube.lsst.codes since it's a private IP address.  This works
on int and stable, because it actually contacts the public address,
which just happens to be the local cluster.  But it doesn't use the
local IP address, unless you refer to it like this.