[DM-27300] Options form plumbing
First, subclass kubespawner to just return all the form options,
and not just profile (which is what the kubespawner code does).

This data is passed to the pre_spawn_hook, which other than validation
is probably a better place to handle all these things (maybe even
the validation there too, just to reduce the number of hooks doing
things).