[DM-26831] Integrate multinamespace kubespawner
First off, we're changing the main.in to point to the upstream
PR for kubespawner that supports multinamespace labs.

Next, stop creating a namespace using the dev-values.yaml, and
change the stop hook to just delete the namespace, since that is
left up to us.  Modify the name of the parameter to make it more
clear.

Also fix up the URL for connecting to the hub to work across
namespaces by adding the namespace as a part of the url.