[DM-28482] User resource creation, pt. 2
Okay so now the user resource creation uses jinja templates.  This
is because of yaml problems with { and }'s, which make it tricky.
(This would make an empty yaml document and then screw up the
templating if you load it.)

So switch to jinja2, and get the auth state and use it for the
templating.