Use an API call to get user authentication data
Rather than relying on the headers set by Gafaelfawr, use the token
provided by Gafaelfawr and make an API call to retrieve the rest of
the user metadata.  Use this to get GIDs as well as group names, and
add them to the user authentication state.

Hopefully fix problems with dependency resolution related to use of
ruamel.yaml.clib on older versions of Python and update all Python
dependencies.

Remove the pinning of tox to Python 3.7 and use the default local
Python instead.