[DM-28482] Allow pod command config
So, the zero-to-jupyterhub chart has the default set to be
jupyterlab-singleuser.  We want it to be None, since that uses
the container's default.  Unforunately, overriding something
to none in helm is a bit tricky, and you can't do it in a
subchart it seems.

Instead of fighting this, just configure it a new way that is
a way we control.