[DM-28950] Dask yaml generation
Okay, so first, fix some whitespace issues I had after switching
to ruamel and fighting with jinja.  Add some spaces between the
{{'s and not.

Use kubespawner to start with the pod manifest for the container
of the lab.  Then we modify it a bit and pass it in as a templated
variable to the resource creation.

The resource creation part has service accounts and such now in
the dev-values.yaml, which are also used by dask.