[DM-28950] Use sanitized pod part
This was confusing since it wasn't really mentioned anywhere, but
there's an attribute map to map camelCase, like how kubernetes does it,
and snake case which is how python does it.  And calling this will
also fix all the null or empty parts of the model too, which is good.
Most of the previous names worked, so that's why it was so confusing.

Also add in the environment variable to let the provisioner know that
we are a dask worker.