[DM-29532] Clean up NubladoConfig
Okay, so here I'm getting rid of the get function, and not inheriting
from LoggingConfigurable.  Getting rid of that dependency lets me make
a reasonable init method which can parse out the yaml.

Then add properties to have strong typing for all the options, rather
than just using the dict everywhere it is used.

Clean up the naming of things as well.