[DM-29532] Hardcode AUTO_REPO_URLS
Nublado2 doesn't really need to know about this, so we shouldn't
pass it through this way.  We should just define this like the
rest of the environment in a static way, unless we need to generate
this based on something at runtime, like the user or environment
or something else.