[DM-29647] Clean up volume mounts
So helm is tricky about merging lists of things.  This causes problems
with the list of volumes and being able to override it.  Same for the
mounts.  You can copy the entire list and add to it, which is kind of
annoying, so I've put the volumes and mounts you need in the base chart,
and made another variable for mounts and volumes that you want to add
to that, which is in the nublado config.  We merge these together
in the hub_config.py.