[DM-29647] Treat user_resources as a template string
Instead of trying to treat user_resources as a yaml document, which
has problems with the jinja templating being in it, let's just treat
it as a static string instead of a yaml document.

Since we're doing that, we don't know it's a list right away, so we
can just template out the whole thing once instead of doing it in a
loop, which is honestly cleaner.

This also allows for us to easily add the annotations and labels
into the user resources as jinja templates.