Add support for Gafaelfawr 2.0
Support a Gafaelfawr token mounted inside the pod and use that by
preference if it exists as the admin token for calls to moneypenny
rather than creating a JWT.

Use the existence of that token as a trigger to use the new token
API for getting user information for a user.  This is a hack, but
it's a temporary hack since we'll be deleting the old code once
Gafaelfawr 2.0 is deployed everywhere.