Accept user resource templates as dicts
In order to allow selective overrides of parts of the user
resource templates via values files in Phalanx, we want to configure
and store them in the Nublado ConfigMap as dicts.  This requires
some additional logic to convert them to text before running them
through Jinja2 for templating (which avoids having to iterate
through every leaf value to see if it needs to be templated).

Switch to the new ruaml.yaml API, since the old one is deprecated.

Detect whether a given resource is a custom resource or not, rather
than forcing the caller to separate the resources into two different
configuration options.