Load Kubernetes configuration earlier
The Kubernetes configuration has to be loaded before the API client
is created, and since we create the API client as a global class
variable during module load, the configuration also has to be loaded
during module load.

This requires unfortunate machinations in the test suite to stub out
the Kubernetes config loader before it is run.