[DM-30180] Use extra_labels instead of common_labels
So, common labels are handled a bit weird, and I was confused.
In the zero to jupyterhub chart, there's some code that takes
the common labels and makes them from the helm data.

Extralabels, on the other hand, is a different property that
is being set, and we should use that instead.  If we ever want
to get to configuring the labels differently for different
resources, then we'll need to add that part of the templating
back in the resource string, but until then, this should be fine.