Wait for a service account token before spawning
We are creating a service account and then spawning a pod using
that service account.  It can take Kubernetes some time to create
the corresponding service account token, and if we attempt to
spawn the pod before that happens, Kubernetes will object.

Wait for Kubernetes to create the service account token before
spawning the pod.  Write this code in a way that can be
transplanted into the upstream KubeSpawner class once we confirm
that this works correctly.