Defer getting Kubernetes clients
Stashing Kubernetes clients as object attributes causes problems
when setting hooks on the spawner, because the Jupyter traitlets
library tries to clone the object and can't clone the Kubernetes
client.  Avoid this by delaying fetching a client to as late as
possible and use the shared_client utility function from inside
the methods that need the client.  This is a bit less efficient
but should avoid the problem.