Remove some unnecessary debugging
There was some debugging logging that was just dumping configuration
data or the JupyterHub auth state.  Remove that to reduce the noise
and copies of secrets logged at the debug level, since it's unlikely
we'll need this debugging unless we're iterating on new development.

This doesn't eliminate logging of secrets; logging at debug level is
not safe to do if the logs are not secure.