Add exponential backoff to namespace deletion
We saw one error on NCSA stable where the connection to etcserver
timed out during namespace creation and then attempted namespace
deletion when spawning a pod.  JupyterHub then gave up and left
the pod in pending deletion state and could not recover without
database surgery.

An ideal solution would be to add a retry loop around all resource
creation, but this is difficult to do given how the code is
currently constructed.  However, it looks like we would have
avoided that problem with a retry loop around namespace deletion
as long as etcserver recovered within the timeout (which it looks
like it did).  We already had all the code available to do that,
so call it from the post_stop hook.