Make spawning robust against an empty cached list
If the cached image list was empty, image_list would not be set in
the submission form, which caused spawning to fail with an exception.
Be robust in the form processing code against this case, and also
modify the HTML template to select "Select uncached image" if the
cached image list were empty.