Further improve spawner JavaScript
Use a different mechanism to select the image_list sentinel value
for the dropdown, since the current method didn't work for me from
the Chrome JavaScript console.  Avoid potentially invalid IDs from
using the display name of images and instead generate unique IDs
from the loop index.  Use onclick instead of onchange for the
dropdown so that selecting the already-selected element still
changes the radio button.

We would prefer to use <label> but JupyterHub styling for <label>
looks awful and it doesn't support the normal Bootstrap way to
override the styling.