[DM-31947] Stop overriding entrypoint
So, in the newer helm chart, it mounts the jupyterhub_config.py
in /usr/local now instead of /etc.  And it fixed whatever issue
with the entrypoint which was the reason I overrode it in the first
place.  So now we can get rid of this.  If we don't, then jupyterhub
ends up confused since both get passed in somehow and then it errors
out.