Do not try to refresh creds for internal requests
The change to force a credential refresh to pick up new Gafaelfawr
credentials before spawn broken authentication from JupyterLab to
JupyterHub because those requests don't include a Gafaelfawr token.

The ingress will ensure that all user requests contain a token and
will check authorization, so it's safe to refresh credentials only
if there is a token and it has changed, and otherwise allow
JupyterHub to use its normal logic.