Update CI, Docker, and dependencies
Reflect changes to the templates by updating the GitHub CI
configuration, the Docker build configuration, and the dependency
declarations.  This will allow CI runs on pull requests, but there
should otherwise not be any significant functionality changes.

Restore explicit pin of jupyterhub to 1.5.0, which apparently was
what was in the last release, and update the Dockerfile base image
to match.