Derive cachemachine URL from base URL
Rather than have the cachemachine URL be a separate configuration
option, derive it from the base URL the same way that the moneypenny
URL and Gafaelfawr URL are handled.  This keeps things more
consistent and will eliminate a use of the cluster-internal
cachemachine URL (which is going away so that a network policy can
be applied).