Stop waiting for ServiceAccount token creation
Kubernetes 1.24 no longer spawns a child Secret for a ServiceAccount,
so we cannot wait for that resource when spawning.  The good news
is that this hopefully also means that the sequencing problem that
we were working around by waiting for it no longer exists.

This will reintroduce the sequencing issue for deployments that
have not yet been upgraded to 1.24.  Thankfully it was relatively
rare.