Only pass groups with GIDs to moneypenny
moneypenny doesn't support groups without GIDs in its model, so it
was failing with the new support for groups from OIDC claims that
have no GIDs (such as at IN2P3-CC).  Only pass groups with GIDs to
moneypenny since it doesn't do anything with the ones without GIDs
anyway.  This will be fixed properly by the new lab controller.

Note that the provisioning flow also doesn't support the user's GID
properly and instead assumes that their primary GID matches their
UID.  This is not fixed in this commit since it should be addressed
by the lab controller, which is hopefully coming soon enough that it
doesn't need to be fixed here (which would require changes to the
moneypenny model and its provisioning Docker image).